requirements.txt
pyTelegramBotAPI==4.14.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
jinja2==3.1.4
aiohttp==3.9.1
tiktoken==0.12.0
pytz==2023.3
numpy==2.4.2 
scikit-learn==1.3.0
